Version code name update
It was called "Zombie" because the project was not being actively maintained. Before that, it was called "Mintaka" since that is a star system in the constellation of Orion, and because "TShock 4.3.21 for Terraria 1.3.4.4 on OTAPI 2.0.0.9 with compatibility for the Terraria Server API version 1.26." is a stupidly long name. Bellatrix is also a star in the constellation of Orion and seems fitting for an actively maintained mobile compatible version of the main PC repository (TShock 4.3.26 for Terraria mobile 1.3.0.7.7 on OTAPI 2.0.0.12).